requirements.txt
pathway-sdk>=0.1.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
aiokafka>=1.8.0
langchain>=0.0.200
openai>=0.27.0